requirements.txt
langchain-tavily
langchain-core
beautifulsoup4==4.13.4
langchain==0.3.27
langgraph==0.6.6
langchainhub==0.1.21
langchain-community==0.3.27
tavily-python==0.7.11
langchain-chroma==0.2.5
python-dotenv==1.1.1
pytest==8.4.1
langchain-openai==0.3.30


#requirements need to update version if its necesssary